project-root/requirements.txt
apache-airflow-providers-postgres
pandas
sqlalchemy
psycopg2-binary
scikit-learn
mlflow
Faker
numpy
